requirements.txt
xlrd==1.1.0
httplib2==0.10.3
pymysql==0.7.11
Flask==0.12.2
requests==2.9.1
gunicorn==19.7.1
JPype1==0.6.3